# api/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go DDD 範本專案 API，實作 Clean Architecture 與 Domain-Driven Design 原則",
        "title": "Go DDD Service API",
        "contact": {},
        "version": "1.2.0"
    },
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "description": "使用 email 和密碼登入，回傳 JWT token pair",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Authenticate a credential",
                "parameters": [
                    {
                        "description": "Login payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_blackhorseya_go-ddd_internal_identity_application_dto.LoginInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_blackhorseya_go-ddd_internal_shared_adapter_http_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_blackhorseya_go-ddd_internal_identity_application_dto.AuthOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_blackhorseya_go-ddd_internal_shared_adapter_http_response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_blackhorseya_go-ddd_internal_shared_adapter_http_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_blackhorseya_go-ddd_internal_shared_adapter_http_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/refresh": {
            "post": {
                "description": "使用 refresh token 取得新的 token pair",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh token pair",
                "parameters": [
                    {
                        "description": "Refresh payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_blackhorseya_go-ddd_internal_identity_application_dto.RefreshInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_blackhorseya_go-ddd_internal_shared_adapter_http_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_blackhorseya_go-ddd_internal_identity_application_dto.AuthOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_blackhorseya_go-ddd_internal_shared_adapter_http_response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_blackhorseya_go-ddd_internal_shared_adapter_http_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_blackhorseya_go-ddd_internal_shared_adapter_http_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "description": "建立新的認證憑證",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new credential",
                "parameters": [
                    {
                        "description": "Registration payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_blackhorseya_go-ddd_internal_identity_application_dto.RegisterInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_blackhorseya_go-ddd_internal_shared_adapter_http_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_blackhorseya_go-ddd_internal_identity_application_dto.CredentialOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_blackhorseya_go-ddd_internal_shared_adapter_http_response.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_blackhorseya_go-ddd_internal_shared_adapter_http_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_blackhorseya_go-ddd_internal_shared_adapter_http_response.Response"
                        }
                    }
                }
            }
        },
        "/healthz": {
            "get": {
                "description": "檢查服務是否存活",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Liveness probe",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_blackhorseya_go-ddd_internal_shared_adapter_http_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_shared_adapter_http_handler.HealthStatus"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/readyz": {
            "get": {
                "description": "檢查服務是否準備好接收流量",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Readiness probe",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_blackhorseya_go-ddd_internal_shared_adapter_http_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_shared_adapter_http_handler.HealthStatus"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_blackhorseya_go-ddd_internal_identity_application_dto.AuthOutput": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "credential": {
                    "$ref": "#/definitions/github_com_blackhorseya_go-ddd_internal_identity_application_dto.CredentialOutput"
                },
                "expires_in": {
                    "type": "integer"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "github_com_blackhorseya_go-ddd_internal_identity_application_dto.CredentialOutput": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_blackhorseya_go-ddd_internal_identity_application_dto.LoginInput": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "github_com_blackhorseya_go-ddd_internal_identity_application_dto.RefreshInput": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "github_com_blackhorseya_go-ddd_internal_identity_application_dto.RegisterInput": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "github_com_blackhorseya_go-ddd_internal_shared_adapter_http_response.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_blackhorseya_go-ddd_internal_shared_adapter_http_response.FieldError"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_blackhorseya_go-ddd_internal_shared_adapter_http_response.FieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_blackhorseya_go-ddd_internal_shared_adapter_http_response.Meta": {
            "type": "object",
            "properties": {
                "pagination": {
                    "$ref": "#/definitions/github_com_blackhorseya_go-ddd_internal_shared_adapter_http_response.Pagination"
                },
                "timestamp": {
                    "type": "string"
                },
                "trace_id": {
                    "type": "string"
                }
            }
        },
        "github_com_blackhorseya_go-ddd_internal_shared_adapter_http_response.Pagination": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "github_com_blackhorseya_go-ddd_internal_shared_adapter_http_response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/github_com_blackhorseya_go-ddd_internal_shared_adapter_http_response.Error"
                },
                "meta": {
                    "$ref": "#/definitions/github_com_blackhorseya_go-ddd_internal_shared_adapter_http_response.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "internal_shared_adapter_http_handler.HealthStatus": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "輸入 Bearer token，格式: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}